+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Changes relevant to the users of python-tcod are documented here.
 This project adheres to [Semantic Versioning](https://semver.org/) since version `2.0.0`.
 
 ## [Unreleased]
+### Added
+- You can new use `SDLConsoleRender.atlas` to access the `SDLTilesetAtlas` used to create it.
+  [#121](https://github.com/libtcod/python-tcod/issues/121)
+
 ### Fixed
 - Fixed the parsing of SDL 2.0.22 headers.  Specifically `SDL_FLT_EPSILON`.